adjust rotation rule to automatically rotate after 30 days
